+++ CHANGELOG.md
@@ -12,6 +12,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for mimir in remoteWrite secret creation.
 - Add mimir ingress secret for basic auth creation.
 
+### Fixed
+
+- Fix mimir heartbeat priority.
+
+### Changed
+
+- Do nothing if mimir is disabled to avoid deleting prometheus-meta-operator managed resources.
+
 ## [0.0.3] - 2024-05-24
 
 ### Changed